Update score.html
I made a small update to the scoreboard functionality. Previously, if you accidentally pressed Cmd + R (or refreshed the page), it would reset the entire score. To address this, I’ve modified the code by saving the score into the local storage so that scores will only reset when the user explicitly clicks the "Reset" button. Now, even if we refresh the page, the scores will remain unchanged and persist as expected.